chiselc: Move spiltting to filter module
