+++ chiselc/src/transforms/mod.rs
@@ -1,4 +1,3 @@
 pub mod filter;
-pub mod filter_splitting;
 pub mod find;
 pub mod utils;
